+++ CHANGELOG.md
@@ -13,6 +13,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Bool `NoColor` to disable color output separately from fatih/color.
+- Method `Ctx.ErrorLine() string` returns a single line error message.
 
 
 [1.0.0] - 2017-01-21